further cleanup; move llava-cli into its own file and rename
